// src/app/Shared/Interfaces/enums.ts
export enum RolUsuario {
  Admin = 'Admin',
  Gerente = 'Gerente',
  Empleado = 'Empleado'
}

export enum EstadoSede {
  Activo = 'activo',
  Mantenimiento = 'mantenimiento',
  Inactivo = 'inactivo'
}

export enum TipoCajon {
  Normal = 'normal',
  Electrico = 'electrico',
  Discapacitado = 'discapacitado'
}

export enum EstadoCajon {
  Libre = 'libre',
  Ocupado = 'ocupado',
  Reservado = 'reservado',
  Inactivo = 'inactivo'
}

export enum EstadoTicket {
  Activo = 'activo',
  Completado = 'completado',
  Cancelado = 'cancelado'
}

export enum EstadoReserva {
  Pendiente = 'pendiente',
  Completado = 'completado',
  Cancelado = 'cancelado'
}